add simple Throughput test and necessary classes; improve SingleNodeTestingContainer;  improve ExchangeCoreClusterClient API; add RESET/NOP commands;

+++ src/main/java/exchange/core2/cluster/model/binary/BinaryDataResult.java
@@ -4,6 +4,6 @@ import exchange.core2.cluster.utils.BufferWritable;
 
 public interface BinaryDataResult extends BufferWritable {
 
-    int getResultCode();
+    short getResultCode();
 
 }
